feat: 修改数据库 INTEGER 类型映射为 Long

+++ starter-cli/src/main/java/io/micronaut/starter/cli/feature/linweiyu/entity/util/TypeConverter.java
@@ -39,7 +39,7 @@ public class TypeConverter {
             case Types.FLOAT:
                 return "Float";
             case Types.INTEGER:
-                return "Integer";
+                return "Long";
             // case Types.JAVA_OBJECT:
             // case Types.LONGNVARCHAR:
             // case Types.LONGVARBINARY:
